@@ -33,7 +33,7 @@
         "slack": "https://codeigniterchat.slack.com"
     },
     "require": {
-        "php": "^7.4 || ^8.0",
+        "php": "^8.1",
         "codeigniter4/framework": "^4.0",
         "codeigniter4/translations": "^4.0",
         "twig/twig": "^3.4"
@@ -41,9 +41,12 @@
     "require-dev": {
         "fakerphp/faker": "^1.9",
         "mikey179/vfsstream": "^1.6",
-        "phpunit/phpunit": "^9.1"
+        "phpunit/phpunit": "^10.5.16"
     },
     "autoload": {
+                "psr-4": {
+            "App\\": "app/"
+        },
         "exclude-from-classmap": [
             "**/Database/Migrations/**"
         ]
