@@ -3,7 +3,7 @@
 use CodeIgniter\CLI\CLI;
 
 // The main Exception
-CLI::write('[' . get_class($exception) . ']', 'light_gray', 'red');
+CLI::write('[' . $exception::class . ']', 'light_gray', 'red');
 CLI::write($message);
 CLI::write('at ' . CLI::color(clean_path($exception->getFile()) . ':' . $exception->getLine(), 'green'));
 CLI::newLine();
@@ -14,7 +14,7 @@ while ($prevException = $last->getPrevious()) {
     $last = $prevException;
 
     CLI::write('  Caused by:');
-    CLI::write('  [' . get_class($prevException) . ']', 'red');
+    CLI::write('  [' . $prevException::class . ']', 'red');
     CLI::write('  ' . $prevException->getMessage());
     CLI::write('  at ' . CLI::color(clean_path($prevException->getFile()) . ':' . $prevException->getLine(), 'green'));
     CLI::newLine();
@@ -50,20 +50,11 @@ if (defined('SHOW_DEBUG_BACKTRACE') && SHOW_DEBUG_BACKTRACE) {
             $function .= $padClass . $error['function'];
         }
 
-        $args = implode(', ', array_map(static function ($value) {
-            switch (true) {
-                case is_object($value):
-                    return 'Object(' . get_class($value) . ')';
-
-                case is_array($value):
-                    return count($value) ? '[...]' : '[]';
-
-                case $value === null:
-                    return 'null'; // return the lowercased version
-
-                default:
-                    return var_export($value, true);
-            }
+        $args = implode(', ', array_map(static fn ($value) => match (true) {
+            is_object($value) => 'Object(' . $value::class . ')',
+            is_array($value)  => count($value) ? '[...]' : '[]',
+            $value === null   => 'null', // return the lowercased version
+            default           => var_export($value, true),
         }, array_values($error['args'] ?? [])));
 
         $function .= '(' . $args . ')';
